feat: 处理某些地方64位数作为被除数时的库依赖问题
在x86架构下，64位除法的进行需要动态加载一些操作系统上提供的库函数，
这就使得我们的项目无法处理64位被除数的问题。我引用了Linux上的do_div
函数来处理这种情况。
但是do_div函数是不安全的，它可能会造成除法的溢出

+++ include/alphaz/type.h
@@ -9,9 +9,7 @@ typedef unsigned int    u32;
 typedef unsigned long   u64;
 
 typedef unsigned long   size_t;
-// typedef long long       loff_t;
-/* 这里本应是long long, 但是当前并没有相关的数学库，所以无法完成64位的除法运算 */
-typedef long            loff_t;
+typedef long long       loff_t;
 typedef int             pid_t;
 
 typedef long            ssize_t;
